fix more whitespaces

+++ musik-spieler/scripts/Playlist.cs
@@ -5,16 +5,18 @@ using System.Collections.Immutable;
 
 namespace Musikspieler.Scripts
 {
-    public class Playlist
+    public class Playlist : IPlaylist
     {
-        private readonly List<Song> songs = [];
+        private readonly List<ISong> songs = [];
 
-        public ImmutableArray<Song> GetAllSongs()
+        public int SongCount => songs.Count;
+
+        public ImmutableArray<ISong> GetAllSongs()
         {
             return songs.ToImmutableArray();
         }
 
-        public IEnumerable<Song> GetEnumerable()
+        public IEnumerable<ISong> GetEnumerable()
         {
             for (int i = 0; i < songs.Count; i++)
             {
@@ -22,28 +24,18 @@ namespace Musikspieler.Scripts
             }
         }
 
-        public struct SongsAddedEventArgs
+        public event Action<SongsAddedEventArgs> SongsAdded = delegate { };
+        public event Action<SongsRemovedEventArgs> SongsRemoved = delegate { };
+
+        public ISong this[int index]
         {
-            public int startIndex;
-            public int count;
+            get { return songs[index]; }
         }
 
-        public struct SongsRemovedEventArgs
+        public bool AddSong(ISong song)
         {
-            public int startIndex;
-            public int count;
-        }
-
-        public event Action<SongsAddedEventArgs> SongsAdded;
-        public event Action<SongsRemovedEventArgs> SongsRemoved;
-
-        public Song GetSongAtIndex(int index)
-        {
-            return songs[index];
-        }
-
-        public void AddSong(Song song)
-        {
+            if (song == null)
+                return false;
             songs.Add(song);
             SongsAddedEventArgs args = new()
             {
@@ -51,10 +43,27 @@ namespace Musikspieler.Scripts
                 count = 1,
             };
             SongsAdded?.Invoke(args);
+            return true;
         }
 
-        public void InsertSongAt(Song song, int index)
+        public bool AddSongs(List<ISong> songList)
         {
+            if (songs == null)
+                return false;
+            SongsAddedEventArgs args = new()
+            {
+                startIndex = songs.Count,
+                count = songList.Count,
+            };
+            songs.AddRange(songList);
+            SongsAdded?.Invoke(args);
+            return true;
+        }
+
+        public bool InsertSongAt(ISong song, int index)
+        {
+            if (song == null || index >= SongCount || index < 0)
+                return false;
             songs.Insert(index, song);
             SongsAddedEventArgs args = new()
             {
@@ -62,10 +71,13 @@ namespace Musikspieler.Scripts
                 count = 1,
             };
             SongsAdded?.Invoke(args);
+            return true;
         }
 
-        public void InsertSongsAt(List<Song> songs, int index)
+        public bool InsertSongsAt(List<ISong> songs, int index)
         {
+            if (songs == null || index >= SongCount || index < 0 || songs.Count < 1)
+                return false;
             songs.InsertRange(index, songs);
             SongsAddedEventArgs args = new()
             {
@@ -73,11 +85,14 @@ namespace Musikspieler.Scripts
                 count = songs.Count,
             };
             SongsAdded?.Invoke(args);
+            return true;
         }
 
-        public void RemoveSong(Song song)
+        public bool RemoveSong(ISong song)
         {
             int index = songs.IndexOf(song);
+            if (index < 0)
+                return false;
             songs.RemoveAt(index);
             SongsRemovedEventArgs args = new()
             {
@@ -85,10 +100,13 @@ namespace Musikspieler.Scripts
                 count = 1,
             };
             SongsRemoved?.Invoke(args);
+            return true;
         }
 
-        public void RemoveSongAt(int index)
+        public bool RemoveSongAt(int index)
         {
+            if (index >= SongCount || index < 0)
+                return false;
             songs.RemoveAt(index);
             SongsRemovedEventArgs args = new()
             {
@@ -96,10 +114,13 @@ namespace Musikspieler.Scripts
                 count = 1,
             };
             SongsRemoved?.Invoke(args);
+            return true;
         }
 
-        public void RemoveSongsAt(int startIndex, int count)
+        public bool RemoveSongsAt(int startIndex, int count)
         {
+            if (startIndex + count >= SongCount || startIndex < 0)
+                return false;
             songs.RemoveRange(startIndex, count);
             SongsRemovedEventArgs args = new()
             {
@@ -107,8 +128,36 @@ namespace Musikspieler.Scripts
                 count = count,
             };
             SongsRemoved?.Invoke(args);
+            return true;
         }
+    }
 
-        public int SongCount => songs.Count;
+    public interface IPlaylist
+    {
+        public int SongCount { get; }
+        public ISong this[int index] { get; }
+        public event Action<SongsAddedEventArgs> SongsAdded;
+        public event Action<SongsRemovedEventArgs> SongsRemoved;
+        public ImmutableArray<ISong> GetAllSongs();
+        public IEnumerable<ISong> GetEnumerable();
+        public bool AddSong(ISong song);
+        public bool AddSongs(List<ISong> songs);
+        public bool InsertSongAt(ISong song, int index);
+        public bool InsertSongsAt(List<ISong> songs, int index);
+        public bool RemoveSong(ISong song);
+        public bool RemoveSongAt(int index);
+        public bool RemoveSongsAt(int startIndex, int count);
+    }
+
+    public struct SongsAddedEventArgs
+    {
+        public int startIndex;
+        public int count;
+    }
+
+    public struct SongsRemovedEventArgs
+    {
+        public int startIndex;
+        public int count;
     }
 }
